+++ package.json
@@ -31,7 +31,6 @@
     "babel-loader": "^7.1.1",
     "babel-preset-env": "^1.5.2",
     "babel-preset-react": "^6.24.1",
-    "bootstrap-loader": "^2.1.0",
     "chalk": "^2.0.1",
     "css-loader": "^0.28.4",
     "eslint": "^4.1.1",
