+++ src/components/bootstrap/bootstrap.scss
@@ -1,2 +1,4 @@
+@import '../../../node_modules/bootstrap/scss/bootstrap';
+
 $black: #000;
 $white: #ffffff;
