+++ src/components/bootstrap/bootstrap.tsx
@@ -5,20 +5,6 @@ import {BrowserRouter as Router, Route, Link, Switch} from 'react-router-dom';
 import Navigation from '../navigation/navigation';
 import Footer from '../footer/footer';
 
-// tslint:disable-next-line
-const App = () => {
-  return (
-    <h1>Content</h1>
-  );
-};
-
-// tslint:disable-next-line
-const About = () => {
-  return (
-    <h1>About</h1>
-  );
-};
-
 // tslint:disable-next-line:no-any
 export default class Bootstrap extends React.Component<any, any> {
 
@@ -32,12 +18,8 @@ export default class Bootstrap extends React.Component<any, any> {
 
         <section className='row main'>
           <div className='container'>
-            {/* Nested routes are managed at the component level*/}
-            <Switch>
-              <Route exact path='/' component={App}/>
-              <Route path='/home' component={App}/>
-              <Route path='/about' component={About}/>
-            </Switch>
+            {/* nested routes passed in as children*/}
+            { this.props.children }
           </div>
         </section>
