--- src/components/footer/footer.scss
@@ -1,1 +1,0 @@
-@import '../bootstrap/bootstrap';
