+++ src/vendor.ts
@@ -1,2 +1,3 @@
 import 'react';
 import 'react-dom';
+import 'bootstrap';
